+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   Added `focusPoint` prop to `ImageBlock` component
+
 ## [0.21.7][] - 2020-02-14
 
 ### Added